Move the filter part of the ui module into the core namespace
The widgets get their own namespace next to the form and menu widgets, the
two abstract models join humhub\models, and the asset joins the other core
bundles.

They land in the core rather than in the stream module, even though the
stream filter panel is their only consumer inside the core. The code knows
nothing about streams, and two outside consumers are not streams at all:
the mail inbox sidebar builds on QueryFilter, and the task search page of
tasks extends four of the input widgets. Filing the filters under stream
would have made those depend on it.

Three views are dropped because nothing rendered them: streamTopicPicker,
topicFilterBlock and contentTypeFilterBlock. They were the only place the
filter code reached into the topic and content modules, and no core class
or known module referenced them, literally or through a widget's view
property. WallStreamFilterNavigation builds its topic and content-type
blocks from ordinary PickerFilterInput instances.

humhub.ui.filter.js moves to the core resources directory and FilterAsset
follows the core convention. The JavaScript module id ui.filter is
unchanged.

Every old class name remains as a deprecated subclass until 1.21.
