+++ protected/humhub/assets/FilterAsset.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ *
+ */
+
+namespace humhub\assets;
+
+use humhub\components\assets\AssetBundle;
+use humhub\modules\topic\assets\TopicAsset;
+
+class FilterAsset extends AssetBundle
+{
+    /**
+     * @inheritdoc
+     */
+    public $sourcePath = '@humhub/resources';
+
+    /**
+     * @inheritdoc
+     */
+    public $js = [
+        'js/humhub/humhub.ui.filter.js',
+    ];
+
+    /**
+     * @inheritdoc
+     */
+    public $depends = [
+        TopicAsset::class,
+    ];
+}
