+++ protected/humhub/modules/stream/assets/StreamAsset.php
@@ -12,7 +12,7 @@ use humhub\assets\CoreExtensionAsset;
 use humhub\components\assets\AssetBundle;
 use humhub\modules\content\assets\ContentAsset;
 use humhub\modules\content\assets\ContentContainerAsset;
-use humhub\modules\ui\filter\assets\FilterAsset;
+use humhub\assets\FilterAsset;
 use humhub\modules\user\assets\UserAsset;
 
 /**
