+++ protected/humhub/modules/stream/widgets/WallStreamFilterNavigation.php
@@ -18,10 +18,10 @@ use humhub\modules\stream\models\filters\DateStreamFilter;
 use humhub\modules\stream\models\filters\DefaultStreamFilter;
 use humhub\modules\stream\models\filters\TopicStreamFilter;
 use humhub\modules\topic\widgets\TopicPicker;
-use humhub\modules\ui\filter\widgets\DatePickerFilterInput;
-use humhub\modules\ui\filter\widgets\FilterNavigation;
-use humhub\modules\ui\filter\widgets\PickerFilterInput;
-use humhub\modules\ui\filter\widgets\RadioFilterInput;
+use humhub\widgets\filter\DatePickerFilterInput;
+use humhub\widgets\filter\FilterNavigation;
+use humhub\widgets\filter\PickerFilterInput;
+use humhub\widgets\filter\RadioFilterInput;
 use humhub\modules\user\models\User;
 use humhub\modules\user\widgets\UserPickerField;
 use Yii;
