+++ protected/humhub/modules/stream/widgets/views/wallStreamFilterNavigation.php
@@ -9,7 +9,7 @@
 use humhub\components\View;
 use humhub\helpers\Html;
 use humhub\modules\stream\widgets\WallStreamFilterNavigation;
-use humhub\modules\ui\filter\widgets\FilterPanel;
+use humhub\widgets\filter\FilterPanel;
 use humhub\widgets\bootstrap\Link;
 
 /* @var $this View */
