+++ protected/humhub/modules/ui/filter/models/QueryFilter.php
@@ -2,19 +2,15 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
- *
  */
 
 namespace humhub\modules\ui\filter\models;
 
-use yii\db\ActiveQuery;
-
-abstract class QueryFilter extends Filter
+/**
+ * @deprecated since 1.20, use \humhub\models\filter\QueryFilter instead. Will be removed in 1.21.
+ */
+abstract class QueryFilter extends \humhub\models\filter\QueryFilter
 {
-    /**
-     * @var ActiveQuery
-     */
-    public $query;
 }
