+++ protected/humhub/modules/ui/filter/widgets/DatePickerFilterInput.php
@@ -2,65 +2,15 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2022 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
- *
  */
 
 namespace humhub\modules\ui\filter\widgets;
 
-use DateTime;
-use humhub\widgets\form\DatePicker;
-use Yii;
-
-class DatePickerFilterInput extends FilterInput
+/**
+ * @deprecated since 1.20, use \humhub\widgets\filter\DatePickerFilterInput instead. Will be removed in 1.21.
+ */
+class DatePickerFilterInput extends \humhub\widgets\filter\DatePickerFilterInput
 {
-    /**
-     * @inheritdoc
-     */
-    public $type = 'date-picker';
-
-    /**
-     * @var array Options for the DatePicker widget
-     */
-    private $datePickerOptions = [];
-
-    /**
-     * @var string data-action-click handler of the input event
-     */
-    public $changeAction = 'inputChange';
-
-    /**
-     * @inheritdoc
-     */
-    protected function initFromRequest()
-    {
-        $filter = Yii::$app->request->get($this->category);
-        $this->value = $filter;
-
-        if (!isset($this->options['placeholder'])) {
-            $this->options['placeholder'] = Yii::$app->formatter->asDate(new DateTime(), Yii::$app->formatter->dateInputFormat);
-        }
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function run()
-    {
-        $this->prepareOptions();
-
-        return DatePicker::widget($this->datePickerOptions);
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function prepareOptions()
-    {
-        parent::prepareOptions();
-        $this->options['data-action-change'] = $this->changeAction;
-        $this->datePickerOptions['options'] = $this->options;
-        $this->datePickerOptions['value'] = $this->value;
-    }
 }
