+++ protected/humhub/modules/ui/filter/widgets/DropdownFilterInput.php
@@ -2,60 +2,15 @@
 
 /**
  * @link https://www.humhub.org/
- * @copyright Copyright (c) 2020 HumHub GmbH & Co. KG
+ * @copyright Copyright (c) HumHub GmbH & Co. KG
  * @license https://www.humhub.com/licences
  */
 
 namespace humhub\modules\ui\filter\widgets;
 
-use humhub\helpers\Html;
-
 /**
- * Dropdown stream filter input type.
- *
- * @since 1.6
- * @package humhub\modules\ui\filter\widgets
+ * @deprecated since 1.20, use \humhub\widgets\filter\DropdownFilterInput instead. Will be removed in 1.21.
  */
-class DropdownFilterInput extends FilterInput
+class DropdownFilterInput extends \humhub\widgets\filter\DropdownFilterInput
 {
-    /**
-     * @inheritdoc
-     */
-    public $view = 'dropdownInput';
-
-    /**
-     * @inheritdoc
-     */
-    public $type = 'dropdown';
-
-    /**
-     * @var array dropdown selection
-     */
-    public $selection = [];
-
-    /**
-     * @var array dropdown items
-     */
-    public $items = [];
-
-    /**
-     * @inheritdoc
-     */
-    public function prepareOptions()
-    {
-        parent::prepareOptions();
-        $this->options['data-action-change'] = 'inputChange';
-        Html::addCssClass($this->options, 'form-control');
-    }
-
-    /**
-     * @inheritdoc
-     */
-    protected function getWidgetOptions()
-    {
-        return array_merge(parent::getWidgetOptions(), [
-            'selection' => $this->selection,
-            'items' => $this->items,
-        ]);
-    }
 }
