+++ protected/humhub/widgets/filter/FilterPanel.php
@@ -1,0 +1,42 @@
+<?php
+
+/**
+ * @link https://www.humhub.org/
+ * @copyright Copyright (c) 2018 HumHub GmbH & Co. KG
+ * @license https://www.humhub.com/licences
+ *
+ */
+
+namespace humhub\widgets\filter;
+
+use humhub\components\Widget;
+use humhub\libs\Sort;
+
+/**
+ * Renders a single stream filter panel which is part of a [[StreamFilterNavigation]].
+ *
+ * @since 1.3
+ * @see FilterNavigation
+ */
+class FilterPanel extends Widget
+{
+    /**
+     * @var array stream filter block definitions
+     */
+    public $blocks = [];
+
+    public $view = 'filterPanel';
+
+    /**
+     * @inheritdoc
+     */
+    public function run()
+    {
+
+        if (empty($this->blocks)) {
+            return '';
+        }
+
+        return $this->render($this->view, ['blocks' => Sort::sort($this->blocks)]);
+    }
+}
