+++ protected/humhub/widgets/filter/views/checkboxInput.php
