+++ protected/humhub/widgets/filter/views/filterListBlock.php
