+++ protected/humhub/widgets/filter/views/pickerInput.php
